pyproject.toml
[build-system]
requires = ["flit_core >=3.11,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "lmitf"
authors = [{name = "Guohao Zhang", email = "guohao2045@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
dynamic = ["version", "description"]
requires-python = ">=3.10"
keywords = ["ai", "llm", "lvm", "api", "interface", "openai", "large-model"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "openai>=1.0",
    "pandas>=2.3",
    "python-dotenv>=1.0",
    "wasabi>=1.1",
    "pillow>=9.0",
]

[project.urls]
Homepage = "https://github.com/colehank/AI-interface"
Repository = "https://github.com/colehank/AI-interface"
Issues = "https://github.com/colehank/AI-interface/issues"
